test-project: simulate the existence of models for now
Because the build system currently mounts sources/models/src/variant,
we need this directory to exist in our project. When conditional
compilation of the model is no longer taking place, the directory will
no longer need to be mounted in the build's Dockerfile and this .keep
file can be removed.